Performance refactor
salt '*' test.ping...now with 40,000 fewer function calls!
